# openapi.yaml
...
openapi: 3.0.0
info:
  title: auto-create-openclaw Agent Instance API
  version: 4.0.0
  description: |
    Express API for auto-creating OpenClaw agent instances.
    
    ## 認證方式
    本 API 使用 JWT Bearer Token 認證。
    - **用戶端點**（`/api/user/me`）：需要 `user` 或 `admin` 角色的有效 JWT
    - **管理員端點**（其餘 `/api/*`）：需要 `admin` 角色的有效 JWT
    
    JWT 必須附加於 `Authorization: Bearer <token>` Header 中。
    
    ## 兩階段註冊流程
    1. `POST /api/register` — 申請實例，取得 QR Code
    2. `GET /api/register/poll/:id` — 輪詢飛書掃碼結果
    
    ## 錯誤格式
    所有錯誤均為 JSON，含 `error` 欄位：
    ```json
    { "error": "錯誤訊息" }
    ```
servers:
  - url: http://localhost:3210
    description: 本機開發伺服器
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        向 `http://127.0.0.1:3001/api/auth/verify` 驗證的 JWT Token。
        - 用戶端點：需要 `user` 或 `admin` 角色
        - 管理員端點：需要 `admin` 角色
  schemas:
...
    InstanceStatus:
      type: string
      enum:
        - pending_scan       # 等待用戶掃碼
        - pending_activation # 掃碼完成，等待管理員激活
        - running            # 已激活，Gateway 運行中
        - stopped            # 已停止
        - error              # 錯誤狀態
        - denied             # 用戶拒絕授權
        - expired            # QR Code 過期
      description: 實例生命週期狀態

    # ── 實例摘要（列表用）───────────────────────────────────────────────────
    InstanceSummary:
      type: object
      properties:
        id:
          type: integer
          description: 資料庫 ID
          example: 7
        agent_id:
          type: string
          description: Agent 唯一識別碼
          example: "user-zhangsan-a1b2c3"
        user_nickname:
          type: string
          description: 用戶暱稱
          example: "張三"
        bot_nickname:
          type: string
          description: 機器人暱稱
          example: "小幫手"
        port:
          type: integer
          description: 容器分配端口
          example: 34521
        status:
          $ref: '#/components/schemas/InstanceStatus'
        budget:
          type: number
          format: float
          description: API 花費上限（美元）
          example: 50.0
        isRunning:
          type: boolean
          description: Gateway 是否運行中
          example: true
        feishu_app_id:
          type: string
          nullable: true
          description: 飛書 App ID
        feishu_app_secret:
          type: string
          nullable: true
          description: 飛書 App Secret（已遮蔽）
          example: "••••"
        openai_api_key:
          type: string
          nullable: true
          description: OpenAI API Key（已遮蔽，只顯示末 4 碼）
          example: "••••sk-4xR9"
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true

    # ── 單一實例詳情（個人用）───────────────────────────────────────────────
...
    RegisterRequest:
      type: object
      required:
        - userNickname
        - botNickname
      properties:
        userNickname:
          type: string
          description: 用戶暱稱（必填，非空，會作為 agentId 的一部分）
          example: "張三"
          minLength: 1
          maxLength: 50
        botNickname:
          type: string
          description: 機器人暱稱（必填，非空）
          example: "小幫手"
          minLength: 1
          maxLength: 50

    # ── 註冊響應 ───────────────────────────────────────────────────────────
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        id:
          type: integer
          description: 資料庫記錄 ID（用於輪詢）
          example: 7
        agentId:
          type: string
          description: 生成的 Agent ID
          example: "user-zhangsan-a1b2c3"
        qrDataUrl:
          type: string
          format: uri
          description: QR Code 圖片（data URI，base64 PNG）
          example: "data:image/png;base64,..."
        verificationUrl:
          type: string
          format: uri
          description: 飛書授權驗證 URL（15 分鐘有效）
          example: "https://..."
        expireIn:
          type: integer
          description: QR Code 有效期（秒）
          example: 300

    # ── 輪詢響應 ────────────────────────────────────────────────────────────
    PollResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - pending              # 等待中
            - completed            # 完成（已取得飛書憑證）
            - denied              # 用戶拒絕
            - expired             # 過期
          example: "pending"
        feishuAppId:
          type: string
          nullable: true
          description: 飛書 App ID（completed 時附加）
          example: "cli_aaa123456789"
        feishuAppSecret:
          type: string
          nullable: true
          description: 飛書 App Secret（completed 時附加）
          example: "abc123..."
        message:
          type: string
          nullable: true
          description: 附加訊息
          example: "飛書機器人建立成功！等待管理員激活。"

    # ── 設定預算請求 ────────────────────────────────────────────────────────
    SetBudgetRequest:
      type: object
      required:
        - budget
      properties:
        budget:
          type: number
          format: float
          description: 預算上限（必須為正數，美元）
          minimum: 0.01
          example: 50.0

    # ── 啟動/停止/刪除 響應 ────────────────────────────────────────────────
    ContainerOperationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        containerName:
          type: string
          description: Docker 容器名稱
          example: "openclaw-user-zhangsan-a1b2c3"
        containerId:
          type: string
          description: Docker 容器 ID
          example: "abc123def456"
        imageName:
          type: string
          description: Docker 映像名稱
          example: "ghcr.io/..."
        action:
          type: string
          description: 執行的操作
          example: "started"

    # ── 激活響應 ───────────────────────────────────────────────────────────
    ActivateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        status:
          type: string
          example: "running"
        containerName:
          type: string
          example: "openclaw-user-zhangsan-a1b2c3"
        containerId:
          type: string
          example: "abc123def456"
        imageName:
          type: string
          example: "ghcr.io/..."

    # ── 健康檢查響應 ────────────────────────────────────────────────────────
...
    SpendResponse:
      type: object
      properties:
        user_id:
          type: string
          description: LiteLLM user_id（即 agent_id）
          example: "user-zhangsan-a1b2c3"
        totalSpend:
          type: number
          format: float
          description: 累計花費（美元）
          example: 12.34
        statusCode:
          type: integer
          example: 200
        timestamp:
          type: string
          format: date-time
          example: "2026-04-10T18:56:00.000Z"

    # ── 錯誤響應 ───────────────────────────────────────────────────────────
    Error:
      type: object
      properties:
        error:
          type: string
          description: 錯誤訊息
          example: "找不到此實例"
  parameters:
    instanceId:
      name: id
      in: path
      required: true
      description: 實例資料庫 ID
      schema:
        type: integer
        example: 7
...
    userIdQuery:
      name: user_id
      in: query
      required: true
      description: LiteLLM user_id（即 agent_id）
      schema:
        type: string
        example: "user-zhangsan-a1b2c3"
    registrationId:
      name: id
      in: path
      required: true
      description: 註冊記錄的資料庫 ID
      schema:
        type: integer
        example: 7
  responses:
    Unauthorized:
      description: 缺少 Authorization Header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Missing Authorization header"
    Forbidden:
      description: Token 無效或權限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Access denied: Admin role required or invalid token"
    NotFound:
      description: 找不到資源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "找不到此實例"
    BadRequest:
      description: 請求參數無效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "暱稱不可為空"
    ServerError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "伺服器錯誤，請稍後再試"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/register:
    post:
      tags:
        - 公開端點
      summary: 申請新實例（第 1 步）
      description: |
        建立資料庫記錄並啟動飛書掃碼流程。
        成功後回傳 QR Code（base64 PNG data URI），供用戶掃碼授權。
        授權 URL 有效期 300 秒（5 分鐘）。
      operationId: RegisterInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              userNickname: "張三"
              botNickname: "小幫手"
      responses:
        "200":
          description: 成功取得 QR Code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                success: true
                id: 7
                agentId: "user-zhangsan-a1b2c3"
                qrDataUrl: "data:image/png;base64,..."
                verificationUrl: "https://open.feishu.cn/app/..."
                expireIn: 300
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/ServerError'
        "503":
          description: 沒有可用端口
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "沒有可用端口"
  /api/register/poll/{id}:
    get:
      tags:
        - 公開端點
      summary: 輪詢掃碼狀態（第 2 步）
      description: |
        查詢用戶是否已完成飛書 QR Code 授權。
        - `pending`：等待中，需繼續輪詢
        - `completed`：授權成功，附加 `feishuAppId` + `feishuAppSecret`
        - `denied`：用戶拒絕授權
        - `expired`：QR Code 過期，需重新註冊
      operationId: PollRegistration
      parameters:
        - $ref: '#/components/parameters/registrationId'
      responses:
        "200":
          description: 輪詢結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollResponse'
              examples:
                pending:
                  summary: 等待掃碼
                  value:
                    status: "pending"
                completed:
                  summary: 授權成功
                  value:
                    status: "completed"
                    feishuAppId: "cli_aaa123456789"
                    feishuAppSecret: "abc123..."
                    message: "飛書機器人建立成功！等待管理員激活。"
                denied:
                  summary: 用戶拒絕
                  value:
                    status: "denied"
                    message: "用戶拒絕了授權"
                expired:
                  summary: QR 過期
                  value:
                    status: "expired"
                    message: "掃碼已過期，請重新註冊"
        "404":
          $ref: '#/components/responses/NotFound'
        "400":
          description: 缺少 device_code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "缺少 device_code"
        "500":
          $ref: '#/components/responses/ServerError'

  # ── 用戶端點 ─────────────────────────────────────────────────────────────
...
  /api/instances:
    get:
      tags:
        - 管理員端點
      summary: 列出所有實例
      description: |
        取得所有用戶實例列表。
        每筆記錄附加 `isRunning`（Gateway 運行狀態）。
        敏感資訊（`feishu_app_secret`、`openai_api_key`）已遮蔽。
      operationId: ListInstances
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 實例列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstanceSummary'
              example:
                - id: 7
                  agent_id: "user-zhangsan-a1b2c3"
                  user_nickname: "張三"
                  bot_nickname: "小幫手"
                  port: 34521
                  status: "running"
                  budget: 50.0
                  isRunning: true
                  feishu_app_secret: "••••"
                  openai_api_key: "••••sk-4xR9"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/instance/{id}/set-budget:
    post:
      tags:
        - 管理員端點
      summary: 設定實例 API 花費上限
      description: 設定用戶實例的 LiteLLM API 花費上限（美元）。
      operationId: SetInstanceBudget
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetBudgetRequest'
            example:
              budget: 50.0
      responses:
        "200":
          description: 設定成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
              example:
                success: true
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/instance/{id}/activate:
    post:
      tags:
        - 管理員端點
      summary: 佈建並啟動容器
      description: 若用戶尚未設定 OpenAI API Key，自動向 LiteLLM Proxy 申請虛擬 Key（限額由 budget 決定，預設 $20），隨後建立並啟動 Docker 容器。
      operationId: ActivateInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/instanceId'
      responses:
        "200":
          description: 啟動成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivateResponse'
              example:
                success: true
                status: "running"
                containerName: "openclaw-user-zhangsan-a1b2c3"
                containerId: "abc123def456"
                imageName: "ghcr.io/..."
        "400":
          description: 尚未完成飛書掃碼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "尚未完成飛書掃碼"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
  /api/instance/{id}/start:
    post:
      tags:
        - 管理員端點
      summary: 啟動 Gateway
      description: |
        啟動指定實例的 Gateway 容器。
        若容器已存在則重啟，若不存在則建立並啟動。
      operationId: StartInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/instanceId'
      responses:
        "200":
          description: 啟動成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerOperationResponse'
              example:
                success: true
                containerName: "openclaw-user-zhangsan-a1b2c3"
                containerId: "abc123def456"
                imageName: "ghcr.io/..."
                action: "started"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
  /api/instance/{id}/stop:
    post:
      tags:
        - 管理員端點
      summary: 停止 Gateway
      description: 停止指定實例的 Gateway 容器。
      operationId: StopInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/instanceId'
      responses:
        "200":
          description: 停止成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerOperationResponse'
              example:
                success: true
                containerName: "openclaw-user-zhangsan-a1b2c3"
                containerId: "abc123def456"
                imageName: "ghcr.io/..."
                action: "stopped"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /api/instance/{id}/delete:
    post:
      tags:
        - 管理員端點
      summary: 刪除實例
      description: |
        刪除實例：移除 Docker 容器、釋放端口、刪除資料庫記錄。
        **此操作不可恢復。**
      operationId: DeleteInstance
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/instanceId'
      responses:
        "200":
          description: 刪除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContainerOperationResponse'
              example:
                success: true
                containerName: "openclaw-user-zhangsan-a1b2c3"
                containerId: "abc123def456"
                imageName: "ghcr.io/..."
                action: "deleted"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /api/spend:
    get:
      tags:
        - 管理員端點
      summary: 用戶花費查詢
      description: 向 LiteLLM Proxy 查詢指定 user_id（即 agent_id）的累計用量。
      operationId: GetUserSpend
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userIdQuery'
      responses:
        "200":
          description: 花費查詢結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendResponse'
              example:
                user_id: "user-zhangsan-a1b2c3"
                totalSpend: 12.34
                statusCode: 200
                timestamp: "2026-04-10T18:56:00.000Z"
        "400":
          description: 缺少 user_id 參數
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "user_id query parameter is required"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'